chore: split the workload API (#16145)
