+++ airbyte-commons-workload/src/main/kotlin/io/airbyte/workload/repository/domain/WorkloadLabel.kt
