+++ airbyte-commons-workload/src/main/kotlin/io/airbyte/workload/repository/domain/WorkloadQueueStats.kt
