+++ airbyte-workload-api-server/src/main/kotlin/io/airbyte/workload/api/domain/Workload.kt
@@ -6,7 +6,6 @@ package io.airbyte.workload.api.domain
 
 import io.airbyte.config.WorkloadPriority
 import io.airbyte.config.WorkloadType
-import io.airbyte.workload.api.domain.Constants.Companion.DEFAULT_GEOGRAPHY
 import io.swagger.v3.oas.annotations.media.Schema
 import java.time.OffsetDateTime
 import java.util.UUID
@@ -19,7 +18,6 @@ data class Workload(
   var labels: MutableList<WorkloadLabel> = mutableListOf(),
   var inputPayload: String = "",
   var logPath: String = "",
-  var geography: String = DEFAULT_GEOGRAPHY,
   var mutexKey: String? = null,
   var type: WorkloadType = WorkloadType.SYNC,
   var terminationSource: String? = null,
