+++ airbyte-workload-init-container/src/test/kotlin/input/CheckHydrationProcessorTest.kt
@@ -112,7 +112,6 @@ class CheckHydrationProcessorTest {
         labels = listOf(),
         inputPayload = "inputPayload",
         logPath = "logPath",
-        geography = "geography",
         type = WorkloadType.CHECK,
         autoId = UUID.randomUUID(),
       )
