+++ airbyte-workload-init-container/src/test/kotlin/input/DiscoverHydrationProcessorTest.kt
@@ -112,7 +112,6 @@ class DiscoverHydrationProcessorTest {
         labels = listOf(),
         inputPayload = "inputPayload",
         logPath = "logPath",
-        geography = "geography",
         type = WorkloadType.DISCOVER,
         autoId = UUID.randomUUID(),
       )
