+++ airbyte-workload-init-container/src/test/kotlin/input/ReplicationHydrationProcessorTest.kt
@@ -170,7 +170,6 @@ class ReplicationHydrationProcessorTest {
         labels = listOf(),
         inputPayload = "inputPayload",
         logPath = "logPath",
-        geography = "geography",
         type = WorkloadType.SYNC,
         autoId = UUID.randomUUID(),
       )
