+++ airbyte-workload-launcher/src/test/kotlin/ClaimedProcessorTest.kt
@@ -77,7 +77,6 @@ class ClaimedProcessorTest {
             labels = listOf(),
             inputPayload = "payload",
             logPath = "logPath",
-            geography = "US",
             type = WorkloadType.SYNC,
             autoId = UUID.randomUUID(),
           ),
@@ -86,7 +85,6 @@ class ClaimedProcessorTest {
             labels = listOf(),
             inputPayload = "payload",
             logPath = "logPath",
-            geography = "US",
             type = WorkloadType.SYNC,
             autoId = UUID.randomUUID(),
           ),
@@ -131,7 +129,6 @@ class ClaimedProcessorTest {
               labels = listOf(),
               inputPayload = "payload",
               logPath = "logPath",
-              geography = "US",
               type = WorkloadType.SYNC,
               autoId = UUID.randomUUID(),
             ),
@@ -140,7 +137,6 @@ class ClaimedProcessorTest {
               labels = listOf(),
               inputPayload = "payload",
               logPath = "logPath",
-              geography = "US",
               type = WorkloadType.SYNC,
               autoId = UUID.randomUUID(),
             ),
@@ -149,7 +145,6 @@ class ClaimedProcessorTest {
               labels = listOf(),
               inputPayload = "payload",
               logPath = "logPath",
-              geography = "US",
               type = WorkloadType.SYNC,
               autoId = UUID.randomUUID(),
             ),
