+++ settings.gradle.kts
@@ -142,6 +142,7 @@ include(":oss:airbyte-commons-temporal")
 include(":oss:airbyte-commons-temporal-core")
 include(":oss:airbyte-commons-converters")
 include(":oss:airbyte-commons-worker")
+include(":oss:airbyte-commons-workload")
 include(":oss:airbyte-config:config-persistence")
 include(":oss:airbyte-config:config-secrets")
 include(":oss:airbyte-featureflag")
@@ -208,6 +209,7 @@ project(":oss:airbyte-commons-temporal").projectDir = file("airbyte-commons-temp
 project(":oss:airbyte-commons-temporal-core").projectDir = file("airbyte-commons-temporal-core")
 project(":oss:airbyte-commons-converters").projectDir = file("airbyte-commons-converters")
 project(":oss:airbyte-commons-worker").projectDir = file("airbyte-commons-worker")
+project(":oss:airbyte-commons-workload").projectDir = file("airbyte-commons-workload")
 project(":oss:airbyte-config:config-persistence").projectDir = file("airbyte-config/config-persistence")
 project(":oss:airbyte-config:config-secrets").projectDir = file("airbyte-config/config-secrets")
 project(":oss:airbyte-csp-check").projectDir = file("airbyte-csp-check")
